refactor: `NumberDto`삭제 및 `LottoWinNumber` domain 추가
