+++ src/main/java/lotto/model/service/LottoHeadQuarter.java
@@ -1,6 +1,6 @@
 package lotto.model.service;
 
-import lotto.controller.dto.NumberDto;
+import lotto.model.domain.LottoWinNumber;
 import lotto.model.domain.vo.BonusNumber;
 import lotto.model.domain.vo.WinNumber;
 
@@ -9,23 +9,11 @@ public class LottoHeadQuarter {
     public LottoHeadQuarter() {
     }
 
-    public NumberDto pickWinNumber(WinNumber winNumber, BonusNumber bonusNumber) {
-        validateDuplication(winNumber, bonusNumber);
-        return NumberDto.toDto(winNumber, bonusNumber);
+    public LottoWinNumber pickNumber(WinNumber winNumber, BonusNumber bonusNumber) {
+        return LottoWinNumber.from(winNumber, bonusNumber);
     }
 
-    private void validateDuplication(WinNumber winNumber, BonusNumber bonusNumber) {
-        if (isDuplicatedBetweenWinAndBonus(winNumber, bonusNumber)) {
-            throw new IllegalArgumentException("당첨 숫자와 보너스 숫자가 중복하면 안됩니다.");
-        }
-    }
-
-    private boolean isDuplicatedBetweenWinAndBonus(WinNumber winNumber, BonusNumber bonusNumber) {
-        return winNumber.getWinNumber().stream()
-                .anyMatch((w) -> w.equals(bonusNumber.getBonusNumber()));
-    }
-
-    public void drawWinner(Lottos playerLottos, WinNumber winNumber, BonusNumber bonusNumber) {
+    public void drawWinner() {
 
     }
